Register Surrounder as package
